refactor(launcher): finalize modal host modes and split the shared surface
Redesign the host-GPU option as a proper boolean toggle, separate from
the GPU type

- resolve_host_on_gpu resolves MIRUMOJI_HOST_ON_GPU as a bool (flag,
  config, shell, default), and build_host_app reads the GPU type from
  MIRUMOJI_MODAL_GPU on its own, falling back to the default GPU only
  when the host runs on one

- wire --host-on-gpu/--no-host-on-gpu and --nonpreemptible/--preemptible
  into modal deploy, resolving with the same precedence as the other
  deploy options

Split the launcher-only Modal operations (credentials, volumes, URLs,
logs) out of the shared mirumoji.modal module into a new
launcher.modal.lifecycle facade, leaving mirumoji.modal as the
server-shared surface (ensure_deployed, stop, and their helpers)

Fix modal logs --follow so Ctrl+C stops it: the follow now runs through
the same process.stream and StreamHandle infrastructure the docker logs
command uses, so it cancels cleanly and can be driven from the GUI
