+++ apps/mirumoji/src/mirumoji/launcher/core/envfile.py
@@ -18,12 +18,10 @@ from dotenv import dotenv_values
 from ... import __version__
 from .constants import (
     CONFIG_ENV_VAR_NAMES,
-    DEFAULT_HOST_GPU,
     HOST_NONPREEMPTIBLE_VAR,
     HOST_ON_GPU_VAR,
     IMAGE_SOURCE_VAR,
     IMAGE_VERSION_VAR,
-    MODAL_GPU_VAR,
     MODAL_HOST_VARS,
     TRANSCRIBE_BACKEND_VAR,
 )
@@ -364,35 +362,33 @@ def host_config(env: Mapping[str, str]) -> dict[str, str]:
     }
 
 
-def resolve_host_gpu(env_path: Path, flag: str | None = None) -> str | None:
+def resolve_host_on_gpu(env_path: Path, flag: bool | None = None) -> bool:
     """
-    Resolves the GPU a GPU host runs on, or `None` for a CPU host
+    Reports whether the host runs on a GPU instead of offloading to the worker
 
     info: Order of Precedence
         - The explicit `flag` override, when provided
 
-        - The persisted `MIRUMOJI_HOST_ON_GPU` toggle and `MIRUMOJI_MODAL_GPU`
-          type, overlaid with the process environment
+        - The persisted `MIRUMOJI_HOST_ON_GPU`, overlaid with the process
+          environment
 
-        - The default GPU, when the host is on a GPU but no type is set
+        - The default (`False`)
 
     A CPU host (the toggle unset or `0`) offloads transcription to the separate
-    worker, so this returns `None`
+    worker, so this returns `False`
 
     Args:
         env_path (Path): The managed config file to read
-        flag (str | None): An explicit GPU override, or `None` to resolve from
-            the config
+        flag (bool | None): An explicit override, or `None` to resolve from the
+            config
 
     Returns:
-        The GPU type for a GPU host, or `None` for a CPU host
+        `True` when the host should run on a GPU with the in-process backend
     """
     if flag is not None:
-        return flag or None
-    values = overlay_environ(read(env_path), [HOST_ON_GPU_VAR, MODAL_GPU_VAR])
-    if values.get(HOST_ON_GPU_VAR) != "1":
-        return None
-    return values.get(MODAL_GPU_VAR) or DEFAULT_HOST_GPU
+        return flag
+    values = overlay_environ(read(env_path), [HOST_ON_GPU_VAR])
+    return values.get(HOST_ON_GPU_VAR) == "1"
 
 
 def resolve_host_nonpreemptible(
